requirements.txt
pydantic[email]~=2.4.2
fastapi~=0.104.1
numpy~=1.26.1
uvicorn~=0.24.0
motor~=3.3.1
pymongo~=4.6.0
pendulum~=2.1.2
passlib~=1.7.4
pytest~=7.4.3
python-jose[cryptography]~=3.3.0
python-multipart~=0.0.6